Arreglado el login, con los cambios de fondo y estilos
