+++ src/front/styles/index.css
@@ -1,18 +1,38 @@
 /* 
     General Styles used on every website (Don't Repeat Yourself)
 */
-body{
+body {
+    font-family: "Reenie Beanie", cursive;
     margin: 0;
     padding: 0;
-    background-image: url("https://i.postimg.cc/kgntGNgM/montana-glenorchy-en-nueva-zelanda-3840x2160-xtrafondos-com-1.jpg");
+    background-image: url("https://i.ibb.co/hZc9bJH/pizarra.jpg%22");
     background-repeat: no-repeat;
     background-size: cover;
     background-attachment: fixed;
-    font-family: fantasy
+  }
+  
+section{
+    width: 100vw;
+    height: 100vh;
 }
 
-form{
-    margin: 5rem auto;
-    -webkit-text-stroke: 0.5px gray;
-    color: rgba(0, 0, 0, 0.5);
+form {
+    margin: auto auto;
+    color: white(0, 0, 0, 0.5);
+    border: 2px groove white;
+    border-radius: 5%;
+    width: 300px;
+    height: 350px;
+}
+
+.input{
+    border: none;
+}
+
+.input::-webkit-input-placeholder {
+    color: white;
+}
+
+.link{
+    text-decoration: none;
 }
